backend/requirements.txt
# Simplified FastAPI Backend - Essential Dependencies Only
fastapi==0.104.1
uvicorn==0.24.0
PyJWT==2.8.0
# Using databases instead of asyncpg for better Windows compatibility
databases[postgresql]==0.8.0
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
aiohttp==3.9.1